requirements.txt
streamlit
requests
beautifulsoup4
scikit-learn
google-genai



